Moved Location Services to Core
