+++ DeviceControlApp.NunitTests/DeviceControlApp.NunitTests.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="..\packages\NUnit.3.11.0\build\NUnit.props" Condition="Exists('..\packages\NUnit.3.11.0\build\NUnit.props')" />
   <PropertyGroup>
@@ -50,6 +50,10 @@
     <Compile Include="LocationPageTests.cs" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\DeviceControlApp.Core\DeviceControlApp.Core.csproj">
+      <Project>{a91b6bc7-8b9e-4b5c-aefb-2b72ef843a0a}</Project>
+      <Name>DeviceControlApp.Core</Name>
+    </ProjectReference>
     <ProjectReference Include="..\DeviceControlApp\DeviceControlApp.csproj">
       <Project>{F3B95DBC-280E-480C-8F52-1F570B8AB17E}</Project>
       <Name>DeviceControlApp</Name>
