+++ DeviceControlApp.NunitTests/DummyLocationService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
+using DeviceControlApp.Core.Service;
 using DeviceControlApp.Services;
 
 namespace DeviceControlApp.NunitTests
