+++ DeviceControlApp/ViewModel/HomePageViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Windows.Input;
 using DeviceControlApp.Core;
+using DeviceControlApp.Core.Service;
 using DeviceControlApp.Services;
 using DeviceControlApp.View;
 using Xamarin.Forms;
@@ -18,25 +19,9 @@ namespace DeviceControlApp.ViewModel
 
         public HomePageViewModel(IPageService pageService,ILocationService locationService)
         {
-            TestStatic.MyInt = 6;
-            if (TestStatic.MyInt != 6) throw new Exception("Static is not working");
-
-            TestObject testObject1 = new TestObject();
-            testObject1.MyInt = 9;
-
-            TestObject testObject2 = new TestObject();
-            testObject2.MyInt = 10;
-
-            if (testObject1.MyInt != 9) throw new Exception("Object is not working");
-
-
-
             _pageService = pageService;
             _locationService = locationService;
             GoToNextCommand = new RelayCommand(GoToNextPage);
-
-
-
         }
 
         public async void GoToNextPage()
